Update project business logic.

+++ Task3/arkpz-pzpi-22-6-kryvko-roman-task3/Atark/src/main/java/ua/nure/kryvko/roman/Atark/controller/ControllerService.java
@@ -1,7 +1,11 @@
 package ua.nure.kryvko.roman.Atark.controller;
 
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
+import ua.nure.kryvko.roman.Atark.greenhouse.Greenhouse;
 import ua.nure.kryvko.roman.Atark.greenhouse.GreenhouseRepository;
 
 import java.util.List;
@@ -19,10 +23,15 @@ public class ControllerService {
         this.greenhouseRepository = greenhouseRepository;
     }
 
+    @Transactional
     public Controller createController(Controller controller) {
         if (controller.getGreenhouse() == null || !greenhouseRepository.existsById(controller.getGreenhouse().getId())) {
             throw new IllegalArgumentException("Greenhouse must exist to create a controller.");
         }
+
+        Greenhouse owner = greenhouseRepository.findById(controller.getGreenhouse().getId())
+                .orElseThrow(() -> new IllegalArgumentException("Greenhouse must exist to create a controller."));
+        controller.setGreenhouse(owner);
         return controllerRepository.save(controller);
     }
 
@@ -34,7 +43,26 @@ public class ControllerService {
         return controllerRepository.findAll();
     }
 
+    @Transactional
+    public Controller updateController(Integer id, Controller controller) {
+        controller.setId(id);
+        if (!controllerRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Controller not found for update");
+        }
+
+        Greenhouse owner = greenhouseRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("Greenhouse must exist to create a controller."));
+        controller.setGreenhouse(owner);
+
+        controllerRepository.save(controller);
+        return controller;
+    }
+
+    @Transactional
     public void deleteController(Integer id) {
+        if (!controllerRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Controller not found for deletion");
+        }
         controllerRepository.deleteById(id);
     }
 }
