+++ Task3/arkpz-pzpi-22-6-kryvko-roman-task3/Atark/src/main/java/ua/nure/kryvko/roman/Atark/notification/NotificationService.java
@@ -1,8 +1,12 @@
 package ua.nure.kryvko.roman.Atark.notification;
 
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
+import ua.nure.kryvko.roman.Atark.user.User;
 import ua.nure.kryvko.roman.Atark.user.UserRepository;
 
 import java.util.Date;
@@ -20,10 +24,15 @@ public class NotificationService {
         this.userRepository = userRepository;
     }
 
+    @Transactional
     public Notification createNotification(Notification notification) {
         if (notification.getUser() == null || !userRepository.existsById(notification.getUser().getId())) {
             throw new IllegalArgumentException("User must exist to create a notification.");
         }
+
+        User owner = userRepository.findById(notification.getUser().getId())
+                .orElseThrow(() -> new IllegalArgumentException("User must exist to create a notification."));
+        notification.setUser(owner);
 
         if (notification.getTimestamp() == null) {
             notification.setTimestamp(new Date());
@@ -36,19 +45,25 @@ public class NotificationService {
         return notificationRepository.findById(id);
     }
 
+    @Transactional
     public Notification updateNotification(Notification notification) {
         if (notification.getId() == null || !notificationRepository.existsById(notification.getId())) {
-            throw new IllegalArgumentException("Notification ID not found for update.");
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Notification not found for update");
         }
+
+        User owner = userRepository.findById(notification.getUser().getId())
+                .orElseThrow(() -> new IllegalArgumentException("User must exist to create a notification."));
+        notification.setUser(owner);
 
         return notificationRepository.save(notification);
     }
 
+    @Transactional
     public void deleteNotificationById(Integer id) {
-        try {
-            notificationRepository.deleteById(id);
-        } catch (EmptyResultDataAccessException e) {
-            throw new IllegalArgumentException("Notification ID not found for deletion.");
+        if (!notificationRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Notification not found for update");
         }
+
+        notificationRepository.deleteById(id);
     }
 }
